Added CLI option to deploy an appstream file
A CLI flag to deploy an appstream file has been added.
When used, the appstream file under the given path is copied into usr/share/metainfo.
This way, users don't have to create an AppDir and manually add the file to the correct location in order for the AppImage to contain one.

Additionally, a -a shortcut has been added to the AppDir flag.