Added CLI option to ignore the excludelist
A new CLI option (--ignore-excludelist) has been added to linuxdeploy. When using it, the excludelist of the core libraries is ignored and the core libraries that are depended upon are included in the AppDir.
The excludelist updating will only take place if this option is not set (as it's not necessary otherwise).

Additionally, the code quality has been improved.